requirements.txt
python-telegram-bot==21.*
fastapi
uvicorn
openai>=1.40.0